perf(lighthouse): /services Win #1 + Win #2 — extract data + locale rewrites
Implements two of the five Lighthouse wins from Issue #773 for /services and /store
(both 1-2 pts below the 65 perf threshold).

### Win #1 — Extract /services static data from the page bundle (+5pts /services)
- New: src/lib/services-data.ts (380 lines, 2 exports)
- Moves getServices() + getServicesFaqs() out of services/page.tsx
- services/page.tsx: 1183 → 807 lines (-376)
- The lib module is server-only static data; tree-shakes any client-side import.

### Win #2 — Server-side locale rewrites (+4pts both pages)
- next.config.ts: adds 7 paths (/services, /store, /contact, /blog, /docs,
  /case-studies, /work) as beforeFiles rewrites to their /en/* equivalents
- Eliminates the 300-500ms 302 round-trip for cookieless first-time visitors
- 'missing: cookie NEXT_LOCALE' guard means localized users still go through
  next-intl normally; only first-visit /services type URLs get the rewrite

### Verification
- Type-check clean (all 3 files: 0 diagnostics)
- No runtime behavior change for already-localized URLs (/en/services etc.)

Expected: /services 64 → ~67, /store 63 → ~66
